+++ tests/testthat/test-run_mod.R
@@ -1,7 +1,9 @@
 test_that(
   desc = "results are consistent with simulated data",
   code = {
-    skip_on_os(c("windows", "mac"))
+    testthat::announce_snapshot_file("sim-strat-curve-params.csv")
+    testthat::announce_snapshot_file("sim-strat-fitted_residuals.csv")
+    skip_on_os(c("windows", "linux"))
     withr::local_seed(1)
     strat1 <- serocalculator::typhoid_curves_nostrat_100 |>
       sim_case_data(n = 100,
@@ -12,7 +14,7 @@ test_that(
       sim_case_data(n = 100,
                     antigen_isos = "HlyE_IgA") |>
       mutate(strat = "stratum 1")
-    dataset <- bind_rows(strat1, strat2)
+    dataset <- dplyr::bind_rows(strat1, strat2)
     withr::with_seed(
       1,
       code = {
@@ -26,18 +28,19 @@ test_that(
           nmc = 2,
           niter = 2, # Number of iterations
           strat = "strat", # Variable to be stratified
-          include_subs = TRUE
         ) |>
-          suppressWarnings() |>
-          magrittr::use_series("curve_params")
+          suppressWarnings()
         
         results |>
           attributes() |>
-          rlist::list.remove("row.names") |>
+          rlist::list.remove(c("row.names", "fitted_residuals")) |>
           expect_snapshot_value(style = "deparse")
         
         results |>
-          ssdtools:::expect_snapshot_data("sim-strat-curve-params")
+          expect_snapshot_data("sim-strat-curve-params")
+
+        attributes(results)$fitted_residuals |>
+          expect_snapshot_data("sim-strat-fitted_residuals")
         
       }
     )
@@ -47,7 +50,8 @@ test_that(
 test_that(
   desc = "results are consistent with SEES data",
   code = {
-    skip_on_os(c("windows", "mac"))
+    testthat::announce_snapshot_file("strat-curve-params.csv")
+    testthat::announce_snapshot_file("strat-fitted_residuals.csv")
     withr::local_seed(1)
     dataset <- serodynamics::nepal_sees 
 
@@ -60,25 +64,27 @@ test_that(
       nmc = 100,
       niter = 100, # Number of iterations
       strat = "bldculres", # Variable to be stratified
-      include_subs = TRUE
     ) |>
-      suppressWarnings() |>
-      magrittr::use_series("curve_params")
+      suppressWarnings()
 
     results |>
       attributes() |>
-      rlist::list.remove("row.names") |>
+      rlist::list.remove(c("row.names", "fitted_residuals")) |>
       expect_snapshot_value(style = "deparse")
 
     results |>
-      ssdtools:::expect_snapshot_data("strat-curve-params")
+      expect_snapshot_data("strat-curve-params")
+
+    attributes(results)$fitted_residuals |>
+      expect_snapshot_data("strat-fitted_residuals")
   }
 )
 
 test_that(
   desc = "results are consistent with unstratified SEES data",
   code = {
-    skip_on_os(c("windows", "mac"))
+    announce_snapshot_file("nostrat-curve-params.csv")
+    announce_snapshot_file("nostrat-fitted_residuals.csv")
     withr::local_seed(1)
     dataset <- serodynamics::nepal_sees 
 
@@ -91,17 +97,86 @@ test_that(
       nmc = 100,
       niter = 100, # Number of iterations
       strat = NA, # Variable to be stratified
-      include_subs = TRUE
     ) |>
-      suppressWarnings() |>
-      magrittr::use_series("curve_params")
+      suppressWarnings()
 
     results |>
       attributes() |>
-      rlist::list.remove("row.names") |>
+      rlist::list.remove(c("row.names", "fitted_residuals")) |>
       expect_snapshot_value(style = "deparse")
 
     results |>
-      ssdtools:::expect_snapshot_data("nostrat-curve-params")
+      expect_snapshot_data("nostrat-curve-params")
+
+    attributes(results)$fitted_residuals |>
+      expect_snapshot_data("nostrat-fitted_residuals")
+  }
+)
+
+test_that(
+  desc = "results are consistent with unstratified SEES data with jags.post
+  included",
+  code = {
+    announce_snapshot_file("nostrat-curve-params-withpost.csv")
+    skip_on_os(c("windows", "linux"))
+    withr::local_seed(1)
+    dataset <- serodynamics::nepal_sees 
+    
+    results <- run_mod(
+      data = dataset, # The data set input
+      file_mod = serodynamics_example("model.jags"),
+      nchain = 2, # Number of mcmc chains to run
+      nadapt = 10, # Number of adaptations to run
+      nburn = 10, # Number of unrecorded samples before sampling begins
+      nmc = 100,
+      niter = 100, # Number of iterations
+      strat = NA, # Variable to be stratified
+      with_post = TRUE
+    ) |>
+      suppressWarnings()
+    
+    results |>
+      attributes() |>
+      rlist::list.remove(c("row.names", "jags.post", "fitted_residuals")) |>
+      expect_snapshot_value(style = "serialize")
+    
+    results |>
+      expect_snapshot_data("nostrat-curve-params-withpost")
+  }
+)
+
+test_that(
+  desc = "results are consistent with unstratified SEES data with modified 
+  priors",
+  code = {
+    announce_snapshot_file("nostrat-curve-params-specpriors.csv")
+    skip_on_os(c("windows", "linux"))
+    withr::local_seed(1)
+    dataset <- serodynamics::nepal_sees 
+    
+    results <- run_mod(
+      data = dataset, # The data set input
+      file_mod = serodynamics_example("model.jags"),
+      nchain = 2, # Number of mcmc chains to run
+      nadapt = 10, # Number of adaptations to run
+      nburn = 10, # Number of unrecorded samples before sampling begins
+      nmc = 100,
+      niter = 100, # Number of iterations
+      strat = NA, # Variable to be stratified
+      mu_hyp_param = c(1, 4, 1, -3, -1),
+      prec_hyp_param = c(0.01, 0.0001, 0.01, 0.001, 0.01),
+      omega_param = c(1, 20, 1, 10, 1),
+      wishdf_param = 10,
+      prec_logy_hyp_param = c(3, 1)
+    ) |>
+      suppressWarnings()
+    
+    results |>
+      attributes() |>
+      rlist::list.remove(c("row.names", "fitted_residuals")) |>
+      expect_snapshot_value(style = "serialize")
+    
+    results |>
+      expect_snapshot_data("nostrat-curve-params-specpriors")
   }
 )